refactor(efi): extract GUID name table to efi/guid_fmt.rs
Move the ~350-line GUID-to-name lookup table and GuidFmt display wrapper
out of boot_services.rs (2,915 lines) into a dedicated guid_fmt module.
This improves separation of concerns — the lookup table is pure data
with no dependency on boot services internals.

boot_services.rs shrinks from 2,915 to 2,562 lines.
