+++ src/efi/mod.rs
@@ -6,6 +6,7 @@
 pub mod allocator;
 pub mod auth;
 pub mod boot_services;
+pub mod guid_fmt;
 pub mod protocols;
 #[cfg(feature = "rt-log")]
 pub mod rtlog;
